refactor(dashboard): give secretCaption its own module
`secretCaption.test.ts` named a module that did not exist: the function
was exported from `KeysPage.tsx`, and of the eleven `*.test.ts` files
under `src/features` it was the only one importing its subject from a
component. `isVirtualUser` travels with it, since the caption is its
only caller outside the table's own rows.
