+++ web/src/features/keys/KeysPage.tsx
@@ -63,6 +63,7 @@ import {
 } from "@/shared/helpers/tableSelection"
 import { useSelectedWorkspace } from "@/shared/hooks/SelectedWorkspace"
 import { useDeployment } from "@/shared/hooks/useDeployment"
+import { isVirtualUser, secretCaption } from "./secretCaption"
 
 // ---------- helpers ----------
 
@@ -99,44 +100,12 @@ function toDatetimeLocal(iso: string | null): string {
   return `${d.getFullYear()}-${pad(d.getMonth() + 1)}-${pad(d.getDate())}T${pad(d.getHours())}:${pad(d.getMinutes())}`
 }
 
-const isVirtualUser = (userId: string | null): boolean =>
-  (userId ?? "").startsWith("apikey-")
-
 const label = (k: ApiKey): string => k.key_name ?? k.id
 
 // Stable row-key getter so DataTable's per-row cache holds across re-renders.
 const getKeyRowKey = (k: ApiKey): string => k.id
 
 // ---------- the one-time secret ----------
-
-/**
- * The line under the key, naming what was just made.
- *
- * Three facts, each dropped when it is not there rather than printed empty. The
- * owner is a name or nothing: `user_id` is an identifier, and an identifier in a
- * sentence is noise to the person who just chose the owner from a list. There is
- * deliberately no spend figure, though the frame it sits in is where one would
- * look for it: a key has no budget of its own, only its owner's, which is what
- * the paragraph above the table already says and links to.
- */
-export function secretCaption(
-  result: CreateKeyResponse,
-  memberLabels: ReadonlyMap<string, string>,
-): string {
-  const owner =
-    result.user_id && !isVirtualUser(result.user_id)
-      ? memberLabels.get(result.user_id)
-      : undefined
-  return [
-    owner ? `Owner ${owner}` : null,
-    accessLabel(result.allowed_models).text,
-    result.expires_at
-      ? `expires ${formatDate(result.expires_at)}`
-      : "never expires",
-  ]
-    .filter((part) => part !== null)
-    .join(" · ")
-}
 
 /**
  * The plaintext key, shown once, and everything that has to travel with it.
